[lpsolve] fix compilation issue in relation to isnan [with Sascha Wunderlich, David Müller]
The compilation wrapper 'ccc' of lpsolve tries to determine
whether the isnan macro is defined. The test program calls
isnan(0), i.e., with an integer argument. On current Arch Linux,
GCC uses _builtin_isnan, which leads to an error due to the integer
argument and the 'ccc' wrapper assumes that isnan is not defined.
This leads to compilation issues later on.

The fix replaces the isnan(0) test call with isnan(0.0).


git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@11216 bbc10eb1-c90d-0410-af57-cb519fbb1720