PrismCL: Catch unhandled Exceptions instead of delegating to default Java handling
We are now catching unhandled Exceptions and errorAndExit. Previously,
we relied on the Java handling, which would print the stack trace and kill
the main thread.

However, it seems that in the presence of thread pools (e.g., with the JAS
library used by param/exact model checking), an unhandled exception would kill
the main thread but not terminate PRISM. This could be triggered
e.g. with

prism prism-examples/dice/dice.pm -pf 'P>0.5[ F s=7 ]' -exact

which generates an unhandled ClassCastException and previously would make
PRISM hang indefinitely.



git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@11537 bbc10eb1-c90d-0410-af57-cb519fbb1720