Improve initial states generation
The explicit engine explores multiple initial states by
explicitly building generating a list of all possible states.
Actually building the list is very slow and runs out of memory easily.

This solution iterates over the cartesion product of the variable domains.
It introduces a deciated iterable for the product which generates tuples on the fly.

Please note that it requires #flatMap from FunctionalIterator instead of Stream.
The latter is broken in the sense that it preallocates memory for the whole result.
Thus, we would run out of memory as well.