Makefile: C++ compiler flag --std=c++11
We would like to use C++11 features in the future, so tell the compiler
to use C++11.

The lpsolve library relies on the WIN32 macro definition to detect Windows
instead of the usual _WIN32. WIN32 seems not to be defined in mingw/gcc
when standard mode is active, so we define it ourselves.

http://nadeausoftware.com/articles/2012/01/c_c_tip_how_use_compiler_predefined_macros_detect_operating_system#WindowsCygwinnonPOSIXandMinGW



git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@12010 bbc10eb1-c90d-0410-af57-cb519fbb1720