Change the default location of the configuration file.
The location of the PRISM configuration file is now OS-dependent. On UNIX-like systems, following the XDG base directory specification, it is prism.settings in ~/.config (or in $XDG_CONFIG_HOME if that overrides it). On OS X, it is prism.settings in ~/Library/Preferences. On anything else it remains in ~/.prism.

If an existing configuration file exists in ~/.prism, that file is still used regardless of the OS.

Closes #38.