DTMCUniformisedSimple: add deadlock accessor methods
As noted in #48, DTMCUniformisedSimple does not implement the full range of ModelExplicit methods. Trying to use it outside of the areas where it's currently used in PRISM can then lead to NullPointerExceptions etc.

We add here accessors for the deadlock states, passing through the method calls to the underlying CTMC.

Additionally, point to the fully-featured uniformised DTMC object / construction method.