GUI renderers/editors: Move from static allocation to allocation on first use
Currently, most of the Setting classes construct a SettingRenderer and SettingEditor statically, for use in the GUI.
When running PRISM from the command-line (PrismCL), we make sure to run Java in headless mode and generally those calls then don't matter.

If there are problems with the AWT configuration, there can be exceptions even for PrismCL. So, here, we
switch from static allocation on class loading to static allocation on first use of getSettingsEditor() and getSettingsRenderer().