FileSetting: On-first-use allocation of file selector
The FileSetting object stores a FileSelector, which is allocated
using defaultSelector() during construction. It can also be set via setFileSelector().

To avoid constructing a FileSelector (and thus call into the Swing/AWT Java package) in
headless / command-line mode, we allocate the default selector on first use, i.e.,
whenever the selector is null in a call to getFileSelector().