DebugJDD: Inhibit leak checks if there was a fatal error during the run
Previously, if DebugJDD found a fatal problem (deref of a node with 0
refs, ...), it would throw an exception, which would then usually lead
to a huge number of leaked nodes on PRISM termination, cluttering the
log.

Now, we don't bother with leak checkes until all the fatal problems
are dealt with.