sparse.cc: fix build_nd_action_vector() result storage
The generated action vector was stored in the 'actions' member of the
NDSparseMatrix that was the last one that was built (remembered in the
global ndsm pointer variable).

Now, we store it in the 'actions' member of the mdp_ndsm that is
passed in, as intended.

This did not result in problems before because the call to
build_nd_action_vector() was always immediately after the
corresponding NDSparseMatrix was built.