Nailgun: Allow the main class started from the nailgun server to be configured via an environment variable
When starting the Nailgun server via 'prism -ng', we now allow the main class to be served
to be configured via the NG_MAINCLASS environment variable. If not set, defaults to prism.PrismCL, as before.

This allows to use nailgun also with other test main methods, e.g., those in the LTL-to-automata classes.