Sparse engine with -exportstrat: Fix crash during strategy initialization
When converting an MTBDD to an integer vector (as is done to prepare the storage for the computed strategy for Pmax-Until using the sparse engine), the MTBDD has to be zero for non-reachable states, as that could lead to out-of-bounds writes to the array during conversion.

This commit adds the required restriction to reach in prism.NondetModelChecker.

+ test case
+ fixes deref of 'yes' during initialization