PRISM startup scripts: remove support for notification via NOTIFY='yes'
Previously, the PRISM startup scripts on MacOS and Linux would call a
GUI notification tool when the PRISM execution has finished and the
environment variable NOTIFY was set to 'yes'.

This feature was not advertised and probably not used much in
practice.  We remove it because it interferes with an exec-based
mechanism of starting the Java VM for PRISM (see next commit).

It is simple to get similar functionality by using a small wrapper
script or other shell functionality.