ExportIterations: Switch to HTTPS in PRISM URL
The HTML file generated by the iteration export loads Javascript and
CSS resources from www.prismmodelchecker.org. Previously, those URLs
used http as the protocol. If such a generated HTML file was loaded
via an https-connections, current browsers would reasonably block
access to the Javascript, as it's not be loaded over a secure
connection.

As the prismmodelchecker.org site now supports https, we switch to
https for loading those resources.