First version of -exportvector <file> switch for the command-line.
Exports the value of a model checked property for all states as a vector.
If <file> is "stdout", this is displayed to the log instead of a file.
Currently, only works for a single property: if there is more than one,
then the file gets overwritten each time a property is checked.

Also fixes a bug in the underlying mechanism for storing results vectors
in the explicit engine (within Results objects).